refactor(signaling): extract WebtritSignalingServiceDirect base class and reorganize Android source layout
- Add WebtritSignalingServiceDirect to platform_interface as shared base for
  direct WebSocket mode (iOS and Android pushBound)
- WebtritSignalingServiceIos now extends the base directly, removing duplicated logic
- Add WebtritSignalingServiceAndroidDirect extending the base with push-isolate
  handoff via IsolateNameServer (onBeforeStart/onConnected hooks)
- WebtritSignalingServiceAndroid delegates pushBound mode to the direct delegate;
  persistent mode continues to use the FGS hub path unchanged
- Reorganize Android lib/src into fgs/ (hub, isolate, hub_connection_manager) and
  direct/ (signaling_service_android_direct) to reflect mode boundaries
- Mirror test directories: test/hub/ and test/isolate/ moved under test/fgs/
- Update barrel exports and all relative imports accordingly
- Rewrite Android README to describe mode principles without implementation detail
