+++ packages/webtrit_signaling_service/webtrit_signaling_service_android/lib/src/fgs/hub/signaling_hub.dart
@@ -6,7 +6,7 @@ import 'package:logging/logging.dart';
 import 'package:webtrit_signaling/webtrit_signaling.dart';
 import 'package:webtrit_signaling_service_platform_interface/webtrit_signaling_service_platform_interface.dart';
 
-import '../constants.dart';
+import '../../constants.dart';
 import 'signaling_hub_codec.dart';
 import 'signaling_hub_command.dart';
