+++ packages/webtrit_signaling_service/webtrit_signaling_service_android/lib/webtrit_signaling_service_android.dart
@@ -1,8 +1,8 @@
 export 'src/constants.dart';
-export 'src/hub/signaling_hub.dart';
-export 'src/hub/signaling_hub_client.dart';
-export 'src/hub/signaling_hub_codec.dart';
-export 'src/hub/signaling_hub_module.dart';
-export 'src/isolate/entry_point.dart';
-export 'src/isolate/signaling_foreground_isolate_manager.dart';
+export 'src/fgs/hub/signaling_hub.dart';
+export 'src/fgs/hub/signaling_hub_client.dart';
+export 'src/fgs/hub/signaling_hub_codec.dart';
+export 'src/fgs/hub/signaling_hub_module.dart';
+export 'src/fgs/isolate/entry_point.dart';
+export 'src/fgs/isolate/signaling_foreground_isolate_manager.dart';
 export 'src/plugin.dart';
