+++ packages/webtrit_signaling_service/webtrit_signaling_service_android/test/integration/bloc_flow_integration_test.dart
@@ -14,9 +14,9 @@ import 'package:ssl_certificates/ssl_certificates.dart';
 import 'package:webtrit_signaling/webtrit_signaling.dart';
 import 'package:webtrit_signaling_service_platform_interface/webtrit_signaling_service_platform_interface.dart';
 
-import 'package:webtrit_signaling_service_android/src/hub/signaling_hub.dart';
-import 'package:webtrit_signaling_service_android/src/hub/signaling_hub_client.dart';
-import 'package:webtrit_signaling_service_android/src/hub/signaling_hub_module.dart';
+import 'package:webtrit_signaling_service_android/src/fgs/hub/signaling_hub.dart';
+import 'package:webtrit_signaling_service_android/src/fgs/hub/signaling_hub_client.dart';
+import 'package:webtrit_signaling_service_android/src/fgs/hub/signaling_hub_module.dart';
 
 // ---------------------------------------------------------------------------
 // Local SignalingClientFactory typedef (mirrors the deleted per-platform type)
