+++ packages/webtrit_signaling_service/webtrit_signaling_service_platform_interface/lib/webtrit_signaling_service_platform_interface.dart
@@ -6,4 +6,5 @@ export 'src/models/signaling_service_mode.dart';
 export 'src/signaling_event_buffer.dart';
 export 'src/signaling_module_impl.dart';
 export 'src/signaling_request_queue.dart';
+export 'src/signaling_service_direct.dart';
 export 'src/signaling_service_platform.dart';
